Deconstruct App into UploadScreen/RunScreen/DetailScreen
