+++ src/javascript/components/App.jsx
@@ -1,189 +1,43 @@
 import React from 'react';
 
 import { connect } from 'store/store.js'
-import StateSelector from 'store/StateSelector.js';
 
-import DoingWorkSpinner from 'components/DoingWorkSpinner.jsx';
 import MainNavi from 'components/MainNavi.jsx';
-import SplitPane from 'components/lib/SplitPane.jsx'
+import DoingWorkSpinner from 'components/DoingWorkSpinner.jsx';
 
-import UploadMainView from 'components/UploadMainView.jsx';
-import UploadSideBar from 'components/UploadSideBar.jsx';
-
-import RunSideBar from 'components/RunSideBar.jsx';
-import DetailSideBar from 'components/DetailSideBar.jsx';
-
-import SingleRunView from 'components/single/SingleRunView.jsx';
-import SingleDetailView from 'components/single/SingleDetailView.jsx';
-
-import TwoRunsView from 'components/two/TwoRunsView.jsx';
-import TwoRunsSummaryView from 'components/two/TwoRunsSummaryView.jsx';
-import TwoDetailView from 'components/two/TwoDetailView.jsx';
-
-import MultiRunView from 'components/multi/MultiRunView.jsx';
-import MultiRunDetailView from 'components/multi/MultiRunDetailView.jsx';
-
-import PrimaryMetricExtractor from 'models/extractor/PrimaryMetricExtractor.js'
-import SecondaryMetricExtractor from 'models/extractor/SecondaryMetricExtractor.js'
-import BenchmarkBundle from 'models/BenchmarkBundle.js';
-import { parseClassNameFromFullName } from 'functions/parse.js';
-
-import Test from 'components/Test.jsx';
-import Test2 from 'components/Test2.jsx';
+import UploadScreen from 'components/UploadScreen.jsx';
+import RunScreen from 'components/RunScreen.jsx';
+import DetailScreen from 'components/DetailScreen.jsx';
 
 /* eslint react/prop-types: 0 */
-// export default class App extends React.Component {
-class App extends React.Component {
-
-    componentDidUpdate() {
-        // spinner will be shown on react.render(..), see entry.jsx
-        // DoingWorkSpinner.hide();
+const App = ({ initialLoading, benchmarkRuns, detailedBenchmarkBundle }) => {
+    if (initialLoading) {
+        return (<div style={ { position: 'fixed', top: '50%', left: '50%' } }><DoingWorkSpinner /></div>);
     }
 
-    render() {
-        const { initialLoading, benchmarkRuns, runView, selectedMetric, detailedBenchmarkBundle, focusedBundles, stateSelector } = this.props;
-
-        console.log("App...");
-        console.log(benchmarkRuns);
-
-        const benchmarkSelection = stateSelector.benchmarkSelection();
-
-        let mainView;
-        let sideBar;
-        if (initialLoading) {
-            return (<div style={ { position: 'fixed', top: '50%', left: '50%' } }><DoingWorkSpinner /></div>);
-        } else if (benchmarkRuns.length == 0) {
-            mainView = <UploadMainView />;
-            sideBar = <UploadSideBar />;
-        } else {
-            const benchmarkBundles = benchmarkSelection.benchmarkBundles;
-
-            if (detailedBenchmarkBundle) { // Details View
-                const detailBundle = benchmarkSelection.benchmarkBundles.find(bundle => bundle.key === detailedBenchmarkBundle) || new BenchmarkBundle({
-                    key: detailedBenchmarkBundle,
-                    name: parseClassNameFromFullName(detailedBenchmarkBundle),
-                    methodNames: [],
-                    benchmarkMethods: []
-                });
-                const secondaryMetrics = Array.from(detailBundle.allBenchmarks().reduce((aggregate, benchmark) => {
-                    Object.keys(benchmark.secondaryMetrics).forEach(metricKey => aggregate.add(metricKey));
-                    return aggregate;
-                }, new Set()));
-                if (benchmarkSelection.runNames.length == 1) {
-                    mainView = <SingleDetailView runName={ benchmarkSelection.runNames[0] } benchmarkBundle={ detailBundle } secondaryMetrics={ secondaryMetrics } />
-                } else if (benchmarkSelection.runNames.length == 2) {
-                    mainView = <TwoDetailView runNames={ benchmarkSelection.runNames } benchmarkBundle={ detailBundle } secondaryMetrics={ secondaryMetrics } />
-                } else {
-                    mainView = <MultiRunDetailView runNames={ benchmarkSelection.runNames } benchmarkBundle={ detailBundle } secondaryMetrics={ secondaryMetrics } />
-                }
-                sideBar = <DetailSideBar
-                    benchmarkBundle={ detailBundle }
-                    benchmarkBundles={ benchmarkBundles }
-                    secondaryMetrics={ secondaryMetrics }
-                />
-            } else { // Run View
-                const metricType = selectedMetric;
-                const metricExtractor = createMetricExtractor(selectedMetric);
-                const categories = ['Benchmarks'];
-                const activeCategory = 'Benchmarks';
-
-                let filteredBenchmarkBundles = metricType === 'Score' ? benchmarkBundles : benchmarkBundles.filter(benchmarkBundle => benchmarkBundle.allBenchmarks().find(benchmark => metricExtractor.hasMetric(benchmark)));
-                let sideBarBenchmarks = filteredBenchmarkBundles;
-                if (stateSelector.shouldFilterFocusedBenchmarks()) {
-                    filteredBenchmarkBundles = filteredBenchmarkBundles.filter(benchmarkBundle => focusedBundles.has(benchmarkBundle.key));
-                }
-                const metricsSet = new Set(['Score']);
-                filteredBenchmarkBundles.forEach(benchmarkBundle => benchmarkBundle.allBenchmarks().forEach(benchmark => {
-                    Object.keys(benchmark.secondaryMetrics).forEach(metricKey => metricsSet.add(metricKey));
-                }));
-                const metrics = Array.from(metricsSet);
-
-
-                if (benchmarkSelection.runNames.length == 1) {
-                    mainView = <SingleRunView
-                        runName={ benchmarkSelection.runNames[0] }
-                        benchmarkBundles={ filteredBenchmarkBundles }
-                        focusedBundles={ focusedBundles }
-                        metricExtractor={ metricExtractor }
-                    />
-                } else if (benchmarkSelection.runNames.length == 2) {
-                    if (runView === 'Summary') {
-                        sideBarBenchmarks = [];
-                        mainView = <TwoRunsSummaryView
-                            runNames={ benchmarkSelection.runNames }
-                            runIndex={ [0, 1] }
-                            benchmarkBundles={ filteredBenchmarkBundles }
-                            minDeviation={ 5 }
-                            metricExtractor={ metricExtractor }
-                        />
-                    } else {
-                        mainView = <TwoRunsView
-                            runNames={ benchmarkSelection.runNames }
-                            benchmarkBundles={ filteredBenchmarkBundles }
-                            metricExtractor={ metricExtractor }
-                        />
-                    }
-                } else {
-                    if (runView === 'Summary') {
-                        sideBarBenchmarks = [];
-                        const firstBenchmark = benchmarkSelection.runNames.length - 2;
-                        const secondBenchmark = benchmarkSelection.runNames.length - 1;
-                        mainView = <TwoRunsSummaryView
-                            runNames={ benchmarkSelection.runNames }
-                            runIndex={ [firstBenchmark, secondBenchmark] }
-                            benchmarkBundles={ filteredBenchmarkBundles }
-                            minDeviation={ 5 }
-                            metricExtractor={ metricExtractor }
-                        />
-                    } else {
-                        mainView = <MultiRunView
-                            runNames={ benchmarkSelection.runNames }
-                            benchmarkBundles={ filteredBenchmarkBundles }
-                            metricExtractor={ metricExtractor }
-                        />
-                    }
-                }
-                sideBar = <RunSideBar
-                    benchmarkBundles={ sideBarBenchmarks }
-                    metrics={ metrics }
-                    metricExtractor={ metricExtractor }
-                    focusedBenchmarkBundles={ focusedBundles }
-                    categories={ categories }
-                    activeCategory={ activeCategory }
-                />
-            }
+    let screen;
+    if (benchmarkRuns.length == 0) {
+        screen = <UploadScreen />
+    } else {
+        if (detailedBenchmarkBundle) { // Details View
+            screen = <DetailScreen />
+        } else { // Run View
+            screen = <RunScreen />
         }
-
-        return (
-            <div>
-                <MainNavi />
-                <div style={ { paddingBottom: 20 + 'px' } }>
-                    <SplitPane left={ mainView } right={ sideBar } />
-                </div>
-                <div>
-                    <Test />
-                    <Test2 />
-                </div>
-            </div>
-        );
     }
 
+    return (
+        <div>
+            <MainNavi />
+            <div style={ { paddingBottom: 20 + 'px' } }>
+                { screen }
+            </div>
+        </div>
+    );
 }
 
-// export default connect((appState2) => ({ appState2 }))(App)
-//TODO don't change on any state change!!
-export default connect(state => ({
-    initialLoading: state.initialLoading,
-    benchmarkRuns: state.benchmarkRuns,
-    runView: state.runView,
-    selectedMetric: state.selectedMetric,
-    detailedBenchmarkBundle: state.detailedBenchmarkBundle,
-    focusedBundles: state.focusedBundles,
-    stateSelector: new StateSelector(state),
+export default connect(({ initialLoading, benchmarkRuns, detailedBenchmarkBundle }) => ({
+    initialLoading,
+    benchmarkRuns: benchmarkRuns,
+    detailedBenchmarkBundle: detailedBenchmarkBundle,
 }))(App)
-
-function createMetricExtractor(metricType) {
-    return metricType === 'Score' ? new PrimaryMetricExtractor() : new SecondaryMetricExtractor(metricType);
-}
-
-
