+++ src/javascript/components/DoingWorkSpinner.jsx
@@ -1,13 +1,10 @@
 import React from 'react';
-// import PropTypes from 'prop-types';
 import { connect } from 'store/store.js'
 import { blue } from 'functions/colors.js'
 var Spinner = require('react-spinkit');
 
 /* eslint react/prop-types: 0 */
 const DoingWorkSpinner = ({ initialLoading, loading }) => {
-    console.log('spinner: ' + initialLoading + '|' + loading);
-
     if (!initialLoading && !loading) {
         return null;
     }
